+++ packages/color2k/src/chain.test.ts
@@ -1,8 +1,8 @@
-import vary from './vary';
+import chain from './chain';
 import { darken, adjustHue, mix } from './';
 
 test('works', () => {
-  const result = vary('red', [
+  const result = chain('red', [
     [darken, 0.1],
     [adjustHue, 180],
     [mix, 'white', 0.3],
